docs(changelog): implement individual changelog page
