--- .github/workflows/changelog.yml
@@ -1,41 +1,0 @@
-name: Update Changelog RC
-
-on:
-  push:
-    paths:
-      - ".changeset/**"
-    branches:
-      - main
-
-jobs:
-  daily:
-    if: ${{ 'chakra-ui/chakra-ui-docs' == github.repository }}
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@master
-        with:
-          ref: main
-
-      - name: Install
-        uses: ./.github/composite-actions/install
-
-      - name: Setup Git
-        run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "github-actions[bot]@users.noreply.github.com"
-
-      - name: Generate the new changelog
-        run: pnpm changelog:gen
-
-      - name: Commit changelog
-        run: |
-          git add .
-          git diff --staged --quiet || git commit -m "docs(changelog): $GITHUB_SHA"
-
-      - name: Create Pull Request
-        uses: peter-evans/create-pull-request@v3
-        with:
-          delete-branch: true
-          title: "docs: update .changelogrc"
-          branch: docs/changelog-updates
