+++ package.json
@@ -12,7 +12,6 @@
     "former-members:gen": "tsx scripts/get-former-members.ts",
     "search-meta:gen": "tsx scripts/get-search-meta.ts",
     "showcase-data:gen": "tsx scripts/get-showcase-data.ts",
-    "changelog:gen": "tsx scripts/get-changelog.ts",
     "contributing:gen": "tsx scripts/build-contributing-page.ts",
     "start": "next start",
     "cache:clean": "rm -rf .next .contentlayer",
@@ -97,6 +96,7 @@
     "pnpm-deduplicate": "^0.3.1",
     "prettier": "^2.7.1",
     "react-multi-ref": "^1.0.0",
+    "semver": "^7.3.7",
     "sharp": "^0.30.7",
     "tsx": "^3.8.2",
     "unified": "^10.1.2",
